+++ CHANGELOG.md
@@ -2,30 +2,58 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.31](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.30..v0.3.31) - January 27, 2022
+
+### Added
+
+- add endpoint `UserSetEmailMutation` (#165)
+
+- add endpoint `UserSetDescriptionMutation` (#166)
+
+### Fixed
+
+- Fix `inLanguage` and `publisher` in metadata api (#167)
+
+## [0.3.30](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.27..v0.3.30) - December 16, 2021
+
+### Added
+
+- Add `EntitiesQuery`.
+
+### Internal
+
+- Update to Rust 1.57.0.
+
+## [0.3.27](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.26..v0.3.27) - November 7, 2021
+
+### Changed
+
+- `UserDeleteBotsMutation` now also returns md5 hashes of removed users.
+
 ## [0.3.26](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.25..v0.3.26) - November 1, 2021
 
 ### Added
 
-- Add query `UserPotentialSpamQuery`
+- Add `UserPotentialSpamQuery`.
 
-- `UuidQuery`: Add `ExerciseGroup.cohesive`
+### Changed
 
-- `UuidQuery`: Add `TaxonomyTerm.taxonomyId`
+- **`UuidQuery`**. Add `ExerciseGroup.cohesive`
+
+- **`UuidQuery`**. Add `TaxonomyTerm.taxonomyId`
 
 ### Fixed
 
-- Creation of threads shall count towards `comments` count in `UserActivityByTypeQuery`
+- Creation of threads now counts towards `comments` in `UserActivityByTypeQuery`.
 
 ### Internal
 
-- Update to new stable rust version (`1.56.0`)
+- Update to Rust 1.56.0.
 
-- Begin of refactoring of integration tests
-
-- Add support for deploying prereleases
+- Add support for deploying prereleases.
 
 ## [0.3.25](https://github.com/serlo/serlo.org-database-layer/compare/a1d5d8261d84ae546914696c363e92ed83b6a17f..v0.3.25) - October 20, 2021
 
 ### Added
 
-- Add mutation `UserDeleteBotsMutation`
+- Add `UserDeleteBotsMutation`.